move Component to Entity
