+++ Engine/Core/ECS/Entity.cpp
@@ -120,3 +120,35 @@ glm::vec4 Entity::getDirection() {
         return glm::normalize(parentEntity->worldMatrix * transform.getDirection());
     }
 }
+
+void Component::setParent(Entity *parentEntity) {
+    m_parentEntity = parentEntity;
+}
+
+Entity *Component::getParent() const {
+    return m_parentEntity;
+}
+
+Transform &Component::getTransform() const {
+    return m_parentEntity->getTransform();
+}
+
+void Component::setProperty(const char *name, PropertyType type, void *p, float min, float max) {
+    Property prop;
+
+    prop.type = type;
+    prop.p = p;
+    prop.min = min;
+    prop.max = max;
+
+    m_properties[name] = prop;
+}
+
+void Component::setProperty(const char *name, PropertyType type, void *p) {
+    Property prop;
+
+    prop.type = type;
+    prop.p = p;
+
+    m_properties[name] = prop;
+}
